+++ SpatialGDK/Source/SpatialGDKEditorToolbar/Public/SpatialGDKEditorToolbar.h
@@ -98,6 +98,9 @@ private:
 
 	void GenerateSchema(bool bFullScan);
 
+	bool IsSnapshotGenerated() const;
+	bool IsSchemaGenerated() const;
+
 	static void ShowCompileLog();
 
 	TSharedPtr<FUICommandList> PluginCommands;